changed scripts filenames and folder structure. created some common functions for point extractions. script not working yet. #118
