+++ scripts/extractions/patch_extractions/extract_optical.py
@@ -13,7 +13,7 @@ from tqdm import tqdm
 
 from worldcereal.openeo.preprocessing import raw_datacube_S2
 
-from worldcereal.openeo.extract_common import (  # isort: skip
+from worldcereal.openeo.extract import (  # isort: skip
     buffer_geometry,  # isort: skip
     get_job_nb_polygons,  # isort: skip
     upload_geoparquet_artifactory,  # isort: skip
@@ -41,7 +41,7 @@ def create_job_dataframe_s2(
         # start_date = max(start_date, S2_L2A_CATALOGUE_BEGIN_DATE)
         # end_date = min(end_date, datetime.now())
 
-        s2_tile = job.tile.iloc[0]  # Job dataframes are split depending on the
+        s2_tile = job.tile.iloc[0]
         h3_l3_cell = job.h3_l3_cell.iloc[0]
 
         # Convert dates to string format
