+++ scripts/extractions/patch_extractions/extract_worldcereal.py
@@ -33,7 +33,7 @@ from worldcereal.openeo.preprocessing import (
 )
 from worldcereal.utils.geoloader import load_reproject
 
-from worldcereal.openeo.extract_common import (  # isort: skip
+from worldcereal.openeo.extract import (  # isort: skip
     get_job_nb_polygons,  # isort: skip
     pipeline_log,  # isort: skip
     upload_geoparquet_artifactory,  # isort: skip
